+++ R/crosstab.R
@@ -31,23 +31,23 @@ crosstab <- function(...) UseMethod("crosstab")
 crosstab.default <- function(vec1, vec2, percent = "none", show_na = TRUE, ...){
 
   if(! percent %in% c("none", "row", "col", "all")){stop("'percent' must be one of 'none', 'row', 'col', or 'all'")}
+  if(!mode(vec1) %in% c("logical", "numeric", "character", "list") | is.matrix(vec1)){
   
-  if(!mode(vec1) %in% c("logical", "numeric", "character", "list")){
     stop("vec1 must be a vector of type logical, numeric, character, list, or factor")}
-  if(!mode(vec2) %in% c("logical", "numeric", "character","list")){
+  if(!mode(vec2) %in% c("logical", "numeric", "character","list") | is.matrix(vec2)){
     stop("vec2 must be a vector of type logical, numeric, character, list, or factor")}
   if(length(vec1) != length(vec2)){ stop("the two vectors are not the same length")}
 
-  dat <- data.frame(vec1=vec1, vec2=vec2, stringsAsFactors = FALSE)
-
+  dat <- data.frame(vec1 = vec1, vec2 = vec2, stringsAsFactors = FALSE)
   dat_col_names <- names(dat)
 
-  if (is.null(names(vec1))) {
+  if(is.null(names(vec1))) {
     var_name <- deparse(substitute(vec1))
   } else {
     var_name <- dat_col_names[[1]]
   }
-
+  
+  var_name <- gsub("[$]", "_", var_name) # for result to have underscore in name instead of .
 
 
   if(!show_na){
@@ -73,18 +73,19 @@ crosstab.default <- function(vec1, vec2, percent = "none", show_na = TRUE, ...){
       dplyr::mutate(n = n / sum(n, na.rm = TRUE))
   }
 
-  # replace NA with string NA in vec2 to avoid invalid col name after spreading
+  # replace NA with string NA_ in vec2 to avoid invalid col name after spreading
   # if this col is a factor, need to add that level to the factor
   if(is.factor(tabl[[2]])){
-    levels(tabl[[2]]) <- c(levels(tabl[[2]]), "NA")
+    levels(tabl[[2]]) <- c(levels(tabl[[2]]), "NA_")
   }
-  tabl[2][is.na(tabl[2])] <- "NA"
+  tabl[2][is.na(tabl[2])] <- "NA_"
 
   # spread to wide, ungroup() for cleanliness of result, and rename 1st col
   tabl %>%
-    tidyr::spread_(dat_col_names[[2]], "n") %>%
+    tidyr::spread_(dat_col_names[[2]], "n", fill = 0) %>%
     dplyr::ungroup() %>%
-    stats::setNames(., c(var_name, names(.)[-1]))
+    stats::setNames(., c(var_name, names(.)[-1])) %>%
+    data.frame(., check.names = FALSE)
 }
 
 #' @inheritParams crosstab.default
@@ -105,7 +106,7 @@ crosstab.data.frame <- function(.data, ...){
   x <- list()
   x[[deparse(columns[[1]])]] <- .data[,deparse(columns[[1]])]
   x[[deparse(columns[[2]])]] <- .data[,deparse(columns[[2]])]
-  x <- as.data.frame(x)
+  x <- as.data.frame(x, stringsAsFactors = FALSE)
 
   # create args list to use with do.call
   arguments <- list()